+++ R/Animate.R
@@ -1,9 +1,8 @@
 #' Generic Animate Method
 #' 
-#' Animate graph-plot objects from select qdap outputs.
+#' Animate select qdap objects.
 #' 
-#' @param x A qdap object (list) with a animatable graph-plot object (e.g., 
-#' \code{\link[qdap]{discourse_map}}).
+#' @param x An animatable qdap object (e.g., \code{\link[qdap]{discourse_map}}).
 #' @param \ldots Arguments passed to Animate method of other classes.
 #' @export
 #' @seealso \code{\link[qdap]{scores}},