+++ R/utils.R
@@ -18,3 +18,39 @@ gg_color_hue <- function(n) {
     hues <- seq(15, 375, length=n+1)
     hcl(h=hues, l=65, c=100)[1:n]
 }
+
+## Pad number and character vectors
+pad <- function (x, padding = max(nchar(as.character(x))), sort = TRUE, 
+    type = "detect") {
+    poss <- c("detect", "numeric", "character", "d", "s")
+    if (!type %in% poss) 
+        stop("type must be: \"detect\", \"numeric\"\\\"d\" or \"character\"\\\"s\"")
+    Rel <- c(NA, "d", "s", "d", "s")
+    type <- Rel[poss %in% type]
+    if (is.na(type)) {
+        type <- ifelse(is.numeric(x), "d", "s")
+    }
+    x <- sprintf(paste0("%0", padding, type), x)
+    if (sort) {
+        x <- sort(x)
+    }
+    x
+}
+
+
+## capture edges from an igraph object
+## Needs igraph::E
+edge_capture <- function(iobj) {
+
+    data.frame(do.call(rbind, 
+    strsplit(bracketX(capture.output(E(iobj)))[-c(1:2)], " -> ")))
+
+}
+
+## generate a from to based on vector of grouping variables
+from_to <- function(x) {
+
+    data.frame(from=as.character(x), to=c(as.character(x[-1]), 
+        "end"), stringsAsFactors = FALSE)
+
+}